Adición comprobaciones y nuevos tests

+++ controllers/platosController.js
@@ -9,6 +9,13 @@ const guardar_platos = async (req, res) => {
         return res.status(400).json({msg: error.message});
     }
 
+    const existePlato = await Plato.findOne(({ where: { nombre_plato: nombre_plato } }));
+
+    if (existePlato){
+        const error = new Error("El plato ingresado ya existe");
+        return res.status(400).json({msg: error.message});
+    }
+
     try {
         
         //Guardar nuevo plato
@@ -27,26 +34,42 @@ const guardar_platos = async (req, res) => {
 
 const eliminar_plato = async (req, res) => {
     //Validar
+    const { id } = req.params;
+
+    const plato = await Plato.findByPk(id)
+
+    if(!plato){
+        const error = new Error("El plato no existe")
+        return res.status(404).json({msg: error.message});
+    }
+
     try {
-        const { id } = req.params;
         await Plato.destroy({
             where:{
                 id,
             },
         });
-        res.sendStatus(204);
+        res.json({msg: 'El plato ha sido borrado exitosamente'});
     } catch (error) {
-        return res.status(500).json({message:error.message});
+        console.log(error);
     }
 };
 
 const actualizar_plato = async (req, res) => {
+    
+    //Validar
+    const { id } = req.params;
+    const {nombre_plato, descripcion, precio} = req.body
+    
+    const plato = await Plato.findByPk(id)
+
+    if(!plato){
+        const error = new Error("El plato no existe")
+        return res.status(404).json({msg: error.message});
+    }
+
     try {
-        //Validar
-        const { id } = req.params;
-        const {nombre_plato, descripcion, precio} = req.body
-        
-        const plato = await Plato.findByPk(id)
+
         plato.nombre_plato = nombre_plato
         plato.descripcion = descripcion
         plato.precio = precio
@@ -54,8 +77,8 @@ const actualizar_plato = async (req, res) => {
         await plato.save();
 
         res.json(plato);
-    } catch (error) {
-        return res.status(500).json({message: error.message});
+    }catch (error) {
+        console.log(error);
     }
 }
